requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bidict==0.23.1
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.0.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.11.3
PyYAML==6.0.2
redis==5.0.7
rich==14.1.0
rich-toolkit==0.15.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.37.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
